rename options.js to optionsUtil.js
